floyd_requirements.txt
opencv-python
pandas
numpy
scipy
scikit-image
tqdm
tensorflow-gpu>1.0
tflearn
tifffile
keras==2.0.4
